fixed warned codes.
git-svn-id: http://alma-dms/svn/sakura/trunk@269 9a79bd28-2a0b-46f8-beb7-14fc758f186d

+++ libsakura/src/optimized_implementation_factory.cc
@@ -60,7 +60,7 @@ void GetCpuFeature(SimdFeature &simd_feature) {
 
 	static SimdFeature const simd_all_false = { };
 	simd_feature = simd_all_false;
-	for (int i = 0; i < ELEMENTSOF(simd_flags); ++i) {
+	for (size_t i = 0; i < ELEMENTSOF(simd_flags); ++i) {
 		if ((reg.ecx & simd_flags[i].ecx) == simd_flags[i].ecx
 				&& (reg.edx & simd_flags[i].edx) == simd_flags[i].edx) {
 			simd_feature.*simd_flags[i].flag = true;
